fix: 카테고리 도메인에 맞게 수정

+++ src/main/java/com/multi/blogging/multiblogging/category/controller/CategoryController.java
@@ -4,6 +4,7 @@ import com.multi.blogging.multiblogging.base.ApiResponse;
 import com.multi.blogging.multiblogging.category.domain.Category;
 import com.multi.blogging.multiblogging.category.dto.request.CategoryRequestDto;
 import com.multi.blogging.multiblogging.category.dto.response.CategoryResponseDto;
+import com.multi.blogging.multiblogging.category.exception.CategoryAccessPermissionDeniedException;
 import com.multi.blogging.multiblogging.category.service.CategoryService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +14,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/category")
+@RequestMapping("{nickname}/category")
 @RequiredArgsConstructor
 public class CategoryController {
     private final CategoryService categoryService;
@@ -21,32 +22,44 @@ public class CategoryController {
     @PostMapping()
     @ResponseStatus(HttpStatus.CREATED)
     @io.swagger.v3.oas.annotations.responses.ApiResponse(description = "상위 카테고리 작성")
-    public ApiResponse<CategoryResponseDto> writeTopCategory(@Valid @RequestBody CategoryRequestDto requestDto){
-        Category category = categoryService.addTopCategory(requestDto.getTitle());
-        return ApiResponse.createSuccess(CategoryResponseDto.of(category));
+    public ApiResponse<CategoryResponseDto> writeTopCategory(@Valid @RequestBody CategoryRequestDto requestDto, @PathVariable(value = "nickname") String nickname) {
+            if (!categoryService.hasCategoryAccessPermission(nickname)){
+                throw new CategoryAccessPermissionDeniedException();
+            }
+            Category category = categoryService.addTopCategory(requestDto.getTitle());
+            return ApiResponse.createSuccess(CategoryResponseDto.of(category));
     }
 
     @PostMapping("/{parent_id}")
     @ResponseStatus(HttpStatus.CREATED)
-    public ApiResponse<CategoryResponseDto> writeChildCategory(@Valid @RequestBody CategoryRequestDto requestDto, @PathVariable("parent_id")Long parentCategoryId ){
-        Category category = categoryService.addChildCategory(parentCategoryId,requestDto.getTitle());
-        return ApiResponse.createSuccess(CategoryResponseDto.of(category));
+    public ApiResponse<CategoryResponseDto> writeChildCategory(@Valid @RequestBody CategoryRequestDto requestDto, @PathVariable("parent_id") Long parentCategoryId, @PathVariable String nickname) {
+            if (!categoryService.hasCategoryAccessPermission(nickname)){
+                throw new CategoryAccessPermissionDeniedException();
+            }
+            Category category = categoryService.addChildCategory(parentCategoryId, requestDto.getTitle());
+            return ApiResponse.createSuccess(CategoryResponseDto.of(category));
     }
 
     @GetMapping("/all")
-    public ApiResponse<List<CategoryResponseDto>> getMyCategories(){
-        List<Category> categories = categoryService.getMyTopCategories();
+    public ApiResponse<List<CategoryResponseDto>> getTopCategories(@PathVariable String nickname) {
+        List<Category> categories = categoryService.getTopCategories(nickname);
         return ApiResponse.createSuccess(categories.stream().map(CategoryResponseDto::of).toList());
     }
 
     @PatchMapping("/{category_id}")
-    public ApiResponse<CategoryResponseDto> updateCategory(@Valid @RequestBody CategoryRequestDto requestDto, @PathVariable("category_id")Long categoryId){
+    public ApiResponse<CategoryResponseDto> updateCategory(@Valid @RequestBody CategoryRequestDto requestDto, @PathVariable("category_id") Long categoryId, @PathVariable String nickname) {
+        if (!categoryService.hasCategoryAccessPermission(nickname)){
+            throw new CategoryAccessPermissionDeniedException();
+        }
         Category updatedCategory = categoryService.updateCategory(requestDto.getTitle(), categoryId);
         return ApiResponse.createSuccess(CategoryResponseDto.of(updatedCategory));
     }
 
     @DeleteMapping("/{category_id}")
-    public ApiResponse<?> deleteCategory(@PathVariable("category_id")Long categoryId){
+    public ApiResponse<?> deleteCategory(@PathVariable("category_id") Long categoryId, @PathVariable String nickname) {
+        if (!categoryService.hasCategoryAccessPermission(nickname)){
+            throw new CategoryAccessPermissionDeniedException();
+        }
         categoryService.deleteCategory(categoryId);
         return ApiResponse.createSuccessWithNoContent();
     }
